+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.2.6:
 
+* Add a utility class for accessing configuration for "services".
+  Useful for storing configuration in a node or data bags as a
+  poor mans service registry.
 * Update `Chef::SearchBlender.blend_search_results_into_node`
   to use partial search where available.